Carriables, lever, pressure switch
For some reason my rock code wasn't in this repo?  Maybe I forgot to
push, but it's there now.  I also did some clean up in the item
hierarchy to prepare for implementation of all objects.  Also, there is
a nice framework for pressure switches and levers now set up.