Pressure switch interaction with player
I added a trigger collider onto the player prefab so that the pressure
switch can know when the player walks over it.  The normal character
controller doesn't through collision or trigger events, so it had to be
done.  I also added a good interface for activating and deactivating an
array of 'Activatable' objects.  All that needs to be done is inherit
from the Activatable class and implement Activate() and Deactivate() in
the necessary fashion. I'll do the same for levers.