Control prompts added
Press 'P' to display prompts for now (only shows E prompt at the
moment). This is for testing purposes until the following functions:
ActivateInteractPrompt(), ActivateJumpPrompt(), and
ActivateSprintPrompt() are implemented into area of the levels. The
prompts display for 5 seconds and then go away. These should be the only
prompts that are needed for the game, as movement and mouse controls
should be obvious.

I couldn't figure out how to make a command prompt hover around an
object, like a lever or rock. I'll continue trying stuff for that, but
this is sufficient for now.

Also added a scroll bar for the credit section, as there will probably
be a shit load of those to add.

Working on sound now.