Added story scenes
I added the story scenes into the game. They don't do anything yet, but
they're ready for when we get the art stills/narration done. The
comments in the scripts tell you exactly what needs to be done to make
them work correctly. The scenes also already have audio sources, so we
just need to drag the audio clips into them when ready.

I might need to make another one for the scoreboard thing that Marsh
wants, but I'll worry about that later.